+++ src/ScottPlot5/ScottPlot5/Primitives/PieSlice.cs
@@ -2,12 +2,11 @@
 
 public class PieSlice
 {
-    public string? Label { get; set; }
+    public string Label { get => LabelStyle.Text; set => LabelStyle.Text = value; }
     public double Value { get; set; }
     public FillStyle Fill { get; set; } = new();
+    public Label LabelStyle { get; set; } = new() { Alignment = Alignment.MiddleCenter };
     public Color FillColor { get => Fill.Color; set => Fill.Color = value; }
-
-    // TODO: deprecate constructors
 
     public PieSlice() { }