+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Markers: Added `Plot.Add.Markers()` to display a collection of marker positions all using the same style (#3283)
 * Axes: Added `Plot.Axes.Remove()` to allow users to remove additional axes they may have added (#3288) _Thanks Felix_
 * Data Streamer and Data Logger: Renamed `IAxisManager` to `IAxisLimitManager` to disambiguate it from the `AxisManager` class (#3289)
+* Pie: Added support for displaying slice label text above each slice (#3295) _Thanks @sterenas_
 
 ## ScottPlot 5.0.20
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2024-01-21_